Repair partially applied Monobank tables

+++ database/migrations/2026_04_01_000002_create_bank_connection_accounts_table.php
@@ -19,7 +19,7 @@ return new class extends Migration {
                 $table->increments('id');
                 $table->timestamps();
 
-                $table->unsignedInteger('bank_connection_id');
+                $table->unsignedBigInteger('bank_connection_id');
                 $table->string('mono_account_id', 255);
                 $table->string('mono_account_type', 50)->nullable();
                 $table->unsignedInteger('mono_currency_code')->nullable();
